+++ src/main/java/ms/parade/interfaces/common/config/FilterConfig.java
@@ -1,8 +1,10 @@
-package ms.parade.interfaces.common;
+package ms.parade.interfaces.common.config;
 
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
+import ms.parade.interfaces.common.filter.TraceIdFilter;
 
 @Configuration
 public class FilterConfig {
