+++ src/main/java/ms/parade/interfaces/common/config/WebConfig.java
@@ -1,10 +1,11 @@
-package ms.parade.interfaces.common;
+package ms.parade.interfaces.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import lombok.RequiredArgsConstructor;
+import ms.parade.interfaces.common.interceptor.AuthInterceptor;
 
 @Configuration
 @RequiredArgsConstructor
