+++ src/main/java/ms/parade/interfaces/common/dto/MessageResponse.java
@@ -1,4 +1,4 @@
-package ms.parade.interfaces.common;
+package ms.parade.interfaces.common.dto;
 
 public record MessageResponse(String message) {
 }
