+++ src/main/java/ms/parade/interfaces/common/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package ms.parade.interfaces.exception;
+package ms.parade.interfaces.common.exception;
 
 public record ErrorResponse(String errorType, String errorMessage) {
 }
