+++ src/main/java/ms/parade/interfaces/common/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package ms.parade.interfaces.exception;
+package ms.parade.interfaces.common.exception;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import jakarta.persistence.EntityNotFoundException;
-import ms.parade.interfaces.common.TraceIdHolder;
+import ms.parade.interfaces.common.util.TraceIdHolder;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
