+++ src/main/java/ms/parade/interfaces/common/util/TraceIdHolder.java
@@ -1,4 +1,4 @@
-package ms.parade.interfaces.common;
+package ms.parade.interfaces.common.util;
 
 public class TraceIdHolder {
     private static final ThreadLocal<String> traceIdHolder = new ThreadLocal<>();
